Fixed callbacks from pre-maturely collected [win]
Event callbacks should be collected with the target object, on x86
machines on windows they were prematurely collected (unsure why it
didn’t happen on x64 builds), event callbacks are now held on the
target object and collected with the disposal of the target object.
This does not apply to general addEventListener function on Tint
objects but native .NET windows EventHandle’s.